books_service/pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.16.0"
uvicorn = "^0.13.4"
taskipy = "^1.8.1"
Django = "^4.1"
strawberry-graphql = "^0.124.0"
strawberry-graphql-django = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
